+++ src/app/disease-form/disease-form.component.spec.ts
@@ -111,7 +111,6 @@ describe('DiseaseFormComponent unit tests', () => {
       featureFlags: {
         FEATURE_FLAG_NON_NOMINAL_NOTIFICATION: true,
         FEATURE_FLAG_FOLLOW_UP_NOTIFICATION_PORTAL_DISEASE: true,
-        FEATURE_FLAG_ANONYMOUS_NOTIFICATION: true,
       },
       ngxLoggerConfig: {
         level: 5,
@@ -309,7 +308,6 @@ describe('DiseaseFormComponent unit tests', () => {
     it('should keep default NominalNotification6_1 when feature flags are disabled', () => {
       environment.diseaseConfig.featureFlags.FEATURE_FLAG_FOLLOW_UP_NOTIFICATION_PORTAL_DISEASE = false;
       environment.diseaseConfig.featureFlags.FEATURE_FLAG_NON_NOMINAL_NOTIFICATION = false;
-      environment.diseaseConfig.featureFlags.FEATURE_FLAG_ANONYMOUS_NOTIFICATION = false;
       Object.defineProperty(routerSpy, 'url', { value: allowedRoutes['followUp'], configurable: true });
 
       component.ngOnInit();