+++ src/app/legacy/common-utils.ts
@@ -26,8 +26,6 @@ import { CodeDisplay, Gender, PractitionerInfo } from 'src/api/notification';
 export const LOCALE_ID_DE: string = 'de-DE';
 registerLocaleData(localeDe, LOCALE_ID_DE, localeDeExtra);
 
-/*** format: TT.MM.JJJJ ***/
-export const UI_DATE_FORMAT_GER: string = 'TT.MM.JJJJ';
 /*** format: dd.LL.yyyy ***/
 export const UI_LUXON_DATE_FORMAT: string = 'dd.LL.yyyy';
 /*** code: DE ***/
@@ -78,7 +76,7 @@ export const ZIP_INTERNATIONAL_ERROR_MSG: string = 'Die Postleitzahl muss aus mi
 export const ZIP_GERMANY_SHORT_ERROR_MSG: string = 'Die Postleitzahl muss aus 3 Ziffern bestehen';
 export const NUMBER_ERROR_MSG: string = 'Bitte geben Sie eine positive Zahl ein.';
 export const VALUE_DEFAULT_PLACEHOLDER: string = 'Bitte eingeben';
-export const VALUE_DEFUALT_SELECT_PLACEHOLDER = 'Bitte auswählen';
+export const VALUE_DEFAULT_SELECT_PLACEHOLDER = 'Bitte auswählen';
 
 // 'None' is hardcoded and not in the enum because it's only needed for the frontend
 export const ExtendedSalutationEnum = {